+++ src/main/java/com/example/Homework_230809/services/implementations/CalculatorServiceImpl.java
@@ -11,25 +11,25 @@ public class CalculatorServiceImpl implements CalculatorService {
     }
 
     @Override
-    public Integer add(Integer num1, Integer num2) {
-        return num1 + num2;
+    public Long add(Integer num1, Integer num2) {
+        return (long)num1 + num2;
     }
 
     @Override
-    public Integer subtract(Integer num1, Integer num2) {
-        return num1 - num2;
+    public Long subtract(Integer num1, Integer num2) {
+        return (long)num1 - num2;
     }
 
     @Override
-    public Integer multiply(Integer num1, Integer num2) {
-        return num1 * num2;
+    public Long multiply(Integer num1, Integer num2) {
+        return (long)num1 * num2;
     }
 
     @Override
-    public Integer divide(Integer num1, Integer num2) {
+    public Double divide(Integer num1, Integer num2) {
         if (num2 == 0){
             throw new IllegalArgumentException("Oops! Divided by zero!");
         }
-        return num1 / num2;
+        return (double)num1 / num2;
     }
 }
